requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow-gpu==1.3.0
keras==2.0.8
pycocotools
opencv-python
h5py==2.10.0
imgaug
opencv-contrib-python==4.2.0.32
IPython[all]
